🐛 Fixing bugs with text lines, gpt after completing tasks and chest image
Adding a new line to the message after the user has finished opening the chest. Adding lines at the end of gpt prompts to make sure they only respond with the necessary text. Putting chest image behind mouseTrackRegion to not being able to click it while in the wizard chat